# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Chissl Server API
  version: 1.0.0
  description: Public and private APIs for the Chissl server. All endpoints are protected unless explicitly marked public.
servers:
  - url: /
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or API Token
  schemas:
...
    User:
      type: object
      properties:
        username: { type: string }
        email: { type: string, nullable: true }
        display_name: { type: string, nullable: true }
        is_admin: { type: boolean }
      additionalProperties: true
    UserList:
      type: array
      items: { $ref: '#/components/schemas/User' }
...
    UserCreateRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
        is_admin: { type: boolean }
    UserUpdateRequest:
      type: object
      required: [username]
      properties:
        username: { type: string }
        password: { type: string }
        is_admin: { type: boolean }
...
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /user/{username}:
    parameters: [{ name: username, in: path, required: true, schema: { type: string } }]
    get:
      summary: Get user by username
      responses: { '200': { description: OK } }
    delete:
      summary: Delete user
      responses: { '202': { description: Accepted } }
...
  /api/users:
    get:
      summary: List users (API)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
    post:
      summary: Create user (API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201': { description: Created }
    put:
      summary: Update user (API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200': { description: OK }
    delete:
      summary: Delete user (API)
      responses: { '204': { description: No Content } }

  # User self APIs (user)
...